cli: cleanup wizard into after unification with ui
