+++ mirrord/cli/src/main.rs
@@ -357,7 +357,6 @@ mod user_data;
 mod util;
 mod verify_config;
 mod vpn;
-mod wizard;
 mod wsl;
 
 pub(crate) use error::{CliError, CliResult};
@@ -1194,7 +1193,6 @@ fn main() -> miette::Result<()> {
             Commands::Up(args) => up::up_command(*args, watch, &user_data).await?,
             Commands::DbBranches(args) => db_branches_command(*args).await?,
             Commands::Queues(args) => queues::queues_command(*args).await?,
-            Commands::Wizard(args) => wizard::wizard_command(*args, watch, &user_data).await?,
             Commands::Fix(args) => fix::fix_command(args).await?,
             #[cfg(windows)]
             Commands::Attach(args) => {
@@ -1203,7 +1201,10 @@ fn main() -> miette::Result<()> {
             }
             #[cfg(windows)]
             Commands::Pitm(args) => pitm::pitm_command(args)?,
-            Commands::Ui(args) => ui::ui_command(args, "/").await?,
+            Commands::Ui { args, command } => ui::ui_command(*args, command, "/").await?,
+            Commands::Wizard { args, telemetry } => {
+                ui::wizard_command(*args, telemetry, watch, &user_data).await?
+            }
             Commands::Session(args) => session::session_command(*args).await?,
             Commands::Kill(args) => session::kill_command(*args).await?,
             #[cfg(unix)]
